Extract `printAngularInterpolation`
